Fixed GetTimelineCaloriesByDay method in foodsqldao

+++ backend/SampleApi/Controllers/TrackingController.cs
@@ -118,7 +118,7 @@ namespace SampleApi.Controllers
         [Route("/api/[controller]/yearlycals")]
         [HttpGet]
         [Authorize]
-        public IEnumerable<int> GetYearlyCalories()
+        public IEnumerable<decimal> GetYearlyCalories()
         {
             return foodDao.GetTimelineCaloriesByDay(CurrentUser.Id, 365);
         }
@@ -126,14 +126,14 @@ namespace SampleApi.Controllers
         [Route("/api/[controller]/monthlycals")]
         [HttpGet]
         [Authorize]
-        public IEnumerable<int> GetMonthlyCalories()
+        public IEnumerable<decimal> GetMonthlyCalories()
         {
             return foodDao.GetTimelineCaloriesByDay(CurrentUser.Id, 30);
         }
         [Route("/api/[controller]/weeklycals")]
         [HttpGet]
         [Authorize]
-        public IEnumerable<int> GetWeeklyCalories()
+        public IEnumerable<decimal> GetWeeklyCalories()
         {
             return foodDao.GetTimelineCaloriesByDay(CurrentUser.Id, 7);
         }
